Rename widgets and update imports
